+++ src/utils/helpers.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-import { decode } from "js-base64";
+import { Base64, decode } from "js-base64";
 import { MoveFunctionId, MoveStructId } from "../types";
 import { AccountAddress } from "../core/accountAddress";
 import { createObjectAddress } from "../core/account/utils/address";
@@ -122,8 +122,8 @@ export function base64UrlToBytes(base64Url: string): Uint8Array {
   while (base64.length % 4 !== 0) {
     base64 += "=";
   }
-  // Use Buffer to convert base64 to Uint8Array
-  return new Uint8Array(Buffer.from(base64, "base64"));
+  // Use js-base64 to convert base64 to Uint8Array
+  return Base64.toUint8Array(base64);
 }
 
 /**